requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
django-cors-headers==3.10.0
djangorestframework-simplejwt==5.0.0
dateutils==0.6.12
gunicorn==20.1.0